unificate toolchain code. comtinue to toolchain code unification. optimize for python import fix is_apple_clang issue. optimize vec_isa_cmd genarator. optimize CppOptions default args.
