+++ torch/_inductor/cpu_vec_isa.py
@@ -1,3 +1,4 @@
+# mypy: allow-untyped-defs
 import dataclasses
 import functools
 import os
@@ -14,28 +15,6 @@ from torch._inductor import config
 _IS_WINDOWS = sys.platform == "win32"
 
 
-# TODO: Move to cpp_builder, when optimize it.
-def get_compiler_version_info(compiler: str) -> str:
-    SUBPROCESS_DECODE_ARGS = ("oem",) if _IS_WINDOWS else ()
-    env = os.environ.copy()
-    env["LC_ALL"] = "C"  # Don't localize output
-    try:
-        version_string = subprocess.check_output(
-            [compiler, "-v"], stderr=subprocess.STDOUT, env=env
-        ).decode(*SUBPROCESS_DECODE_ARGS)
-    except Exception as e:
-        try:
-            version_string = subprocess.check_output(
-                [compiler, "--version"], stderr=subprocess.STDOUT, env=env
-            ).decode(*SUBPROCESS_DECODE_ARGS)
-        except Exception as e:
-            return ""
-    # Mutiple lines to one line string.
-    version_string = version_string.replace("\r", "_")
-    version_string = version_string.replace("\n", "_")
-    return version_string
-
-
 def _get_isa_dry_compile_fingerprint(isa_flags: str) -> str:
     # ISA dry compile will cost about 1 sec time each startup time.
     # Please check the issue: https://github.com/pytorch/pytorch/issues/100378
@@ -43,9 +22,9 @@ def _get_isa_dry_compile_fingerprint(isa_flags: str) -> str:
     # We just record the compiler version, isa options and pytorch version info,
     # and generated them to output binary hash path.
     # It would optimize and skip compile existing binary.
-    from torch._inductor.cpp_builder import cpp_compiler
+    from torch._inductor.cpp_builder import get_compiler_version_info, get_cpp_compiler
 
-    compiler_info = get_compiler_version_info(cpp_compiler())
+    compiler_info = get_compiler_version_info(get_cpp_compiler())
     torch_version = torch.__version__
     fingerprint = f"{compiler_info}={isa_flags}={torch_version}"
     return fingerprint
